requirements.txt
discord.py==1.6.0
jishaku==1.19.1.200
aiohttp==3.6.2